Re-organise interaction components
