+++ src/Timeline.tsx
@@ -2,11 +2,9 @@ import React, { useEffect, useState } from 'react'
 import { Domain, EventComponentFactory, LaneDisplayMode, TimelineEvent, TimelineLane } from './model'
 import { defaultOrderedZoomLevels, ZoomLevels, ZoomScale, zoomScaleWidth } from './shared/ZoomScale'
 import { scaleBand, scaleLinear } from 'd3-scale'
-import { MouseAwareSvg, SvgCoordinates } from './MouseAwareSvg'
-import { MouseCursor } from './MouseCursor'
 import { GridLines } from './layers/GridLines'
 import { ExpandedMarks } from './layers/ExpandedMarks'
-import { InteractionHandling } from './InteractionHandling'
+import { InteractionHandling, MouseCursor, MouseAwareSvg, SvgCoordinates } from './layers/interaction'
 import { noOp } from './utils'
 import { CollapsedMarks } from './layers/CollapsedMarks'
 import { Trimmer, TrimRange, useTrimming } from './trimmer'
