+++ src/index.tsx
@@ -1,8 +1,6 @@
 export * from './model'
 export * from './Timeline'
 export * from './layers/Marks'
-export * from './InteractionHandling'
-export * from './MouseCursor'
-export * from './MouseAwareSvg'
+export * from './layers/interaction'
 export * from './shared/ZoomScale'
 export { TimelineTheme, createTimelineTheme } from './theme'
