+++ src/layers/interaction/InteractionHandling.tsx
@@ -1,8 +1,18 @@
 import React from 'react'
 import { useEffect, useState } from 'react'
-import { Cursor, Domain } from './model'
-import { noOp } from './utils'
+import { Cursor, Domain } from '../../model'
+import { noOp } from '../../utils'
 import { SvgCoordinates } from './MouseAwareSvg'
+import {
+  Anchored,
+  InteractionMode,
+  interactionModeAnimationInProgress,
+  InteractionModeGrabbing,
+  interactionModeHover,
+  interactionModeNone,
+  TrimHover,
+  TrimNone,
+} from './model'
 
 export interface InteractionHandlingProps {
   width: number
@@ -29,75 +39,6 @@ export interface InteractionHandlingProps {
     setTrimHoverMode: (trimHoverMode: TrimHover | TrimNone) => void
   ) => React.ReactNode
 }
-
-interface InteractionModeNone {
-  type: 'none'
-}
-
-const interactionModeNone: InteractionModeNone = { type: 'none' }
-
-interface InteractionModeHover {
-  type: 'hover'
-}
-
-const interactionModeHover: InteractionModeHover = { type: 'hover' }
-
-interface Anchored {
-  variant: 'anchored'
-  anchorX: number
-}
-
-interface InProgress {
-  variant: 'in progress'
-  anchorX: number
-  currentX: number
-}
-
-interface InteractionModeGrabbing extends Anchored {
-  type: 'grabbing'
-}
-
-interface InteractionModePanning extends Anchored {
-  type: 'panning'
-}
-
-type InteractionModeRubberBand =
-  | (Anchored & Readonly<{ type: 'rubber band' }>)
-  | (InProgress & Readonly<{ type: 'rubber band' }>)
-
-interface InteractionModeAnimationInProgress {
-  type: 'animation in progress'
-}
-
-const interactionModeAnimationInProgress: InteractionModeAnimationInProgress = { type: 'animation in progress' }
-
-export interface TrimNone {
-  variant: 'none'
-}
-
-export interface TrimHover {
-  variant: 'trim hover start' | 'trim hover end'
-}
-
-interface TrimInProgress {
-  variant: 'trim start' | 'trim end' | 'trim pan end'
-}
-
-type InteractionModeTrimHover = TrimHover & Readonly<{ type: 'trim' }>
-
-type InteractionModeTrim =
-  | (TrimNone & Readonly<{ type: 'trim' }>)
-  | InteractionModeTrimHover
-  | (TrimInProgress & Readonly<{ type: 'trim' }>)
-
-export type InteractionMode =
-  | InteractionModeNone
-  | InteractionModeHover
-  | InteractionModeAnimationInProgress
-  | InteractionModePanning
-  | InteractionModeRubberBand
-  | InteractionModeTrim
-  | InteractionModeGrabbing
 
 export const InteractionHandling = ({
   width,
