+++ src/layers/interaction/MouseAwareSvg.tsx
