+++ src/layers/interaction/MouseCursor.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react'
-import { ZoomScale } from './shared/ZoomScale'
+import { ZoomScale } from '../../shared/ZoomScale'
 import makeStyles from '@material-ui/core/styles/makeStyles'
 import { orange } from '@material-ui/core/colors'
-import { Cursor } from './model'
-import { InteractionMode } from './InteractionHandling'
-import CursorLabel from './CursorLabel'
+import { Cursor } from '../../model'
+import { InteractionMode } from '.'
+import { CursorLabel } from './CursorLabel'
 
 const useStyles = makeStyles(() => ({
   cursor: {
