+++ src/trimmer/Trimmer.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { makeStyles } from '@material-ui/core'
 import { ScaleLinear } from 'd3-scale'
-import { TrimHover, TrimNone } from '../InteractionHandling'
+import { TrimHover, TrimNone } from '../layers/interaction'
 import TrimHandle from './TrimHandle'
 import Triangle, { TriangleDirection } from '../shared/Triangle'
 import { useTimelineTheme } from '../theme'
